requirements.txt
jinja2==2.8
tweepy==3.5.0
pillow==3.1.1
numpy==1.10.4
noise==1.2.2


